fix: 7 bugs from self-audit bugreport
Critical:
1. grep_tool.c: shell injection — escape single quotes in pattern/path

Bugs:
2. validation.c: cJSON leak when arguments=NULL — track and free
3. agent.c: API key strdup leaked every turn — free after stream
4. interactive.c: cJSON msg_data leaked on every message — cJSON_Delete
5. session.c: non-thread-safe rand/srand — read /dev/urandom directly

Minor:
10. str.c: integer overflow in str_replace_all — use long long
14. interactive.c: tool args truncated at 200 chars for display
15. interactive.c: pi_log_close() on exit

+++ src/util/str.c
@@ -285,8 +285,8 @@ int str_replace_all(Str *s, const char *old, const char *new_str) {
     if (count == 0) return 0;
 
     // Calculate size difference
-    int size_diff = (int)(new_len - old_len);
-    size_t new_total_len = s->len + (size_t)(size_diff * count);
+    long long size_diff = (long long)new_len - (long long)old_len;
+    size_t new_total_len = (size_t)((long long)s->len + size_diff * count);
 
     if (size_diff > 0) {
         // Growing: need more space
